:recycle: Rename package value to raw
