+++ parser/code.go
@@ -1,0 +1,5 @@
+package parser
+
+type Code interface {
+}
+type CodeImpl struct{}
