+++ raw/code.go
@@ -1,0 +1,3 @@
+package raw
+
+type Code struct{}
