--- value/ast.go
@@ -1,3 +1,0 @@
-package value
-
-type Ast struct{}
